create  an entry point for the vnu validator
